engine: declare PFN_pEngineEntry instead of the function declaration
This makes more sense, because pEngineEntry is called only once in the
runner's code. But previously we had our own typedef of pEngineInit in
the runner.

Because of this change, I maked the function `extern` and moved the PAPI
specifier from the header into the .c file to make it visible externally.