Merge Vulkan renderer into the engine
Previously it was a separate library. First I thought it is the
superior way of doing things, but as I started developing it, I
understood that I was completely wrong.

This patch removes libpogona_vulkan_ren target. It also means that
the engine does not need to dynamically load the library. So it
doesn't depend on libdl anymore. Thus, it's more cross-platform,
because we haven't got a way to dynamically load shared objects for
any OS.