Make Vulkan renderer a separate library, again
Everyone makes bad choices, ok?
I found a way to make it a separate library AND be easy and comfortable
to extend. There's no glue between the engine and the renderer now! (I'm
talking about the gEngine thing). It was awful before (which is why I
removed it then).

The renderer is now linked with the engine library. This makes it possible
to use engine's functions without glue.